fixes in blances

+++ Dawem.Models/Response/Requests/Vacations/EmployeeGetVacationsInformationsResponseDTO.cs
@@ -2,7 +2,7 @@
 {
     public class EmployeeGetVacationsInformationsResponseDTO
     {
-        public int VacationsBalance { get; set; }
+        public float VacationsBalance { get; set; }
         public int AcceptedCount { get; set; }
         public int RejectedCount { get; set; }
         public int PendingCount { get; set; }
